Criado codigo especializado para db, server e mudanças no client
