+++ .gitignore
@@ -1,3 +1,4 @@
+*.env
 
 # Created by https://www.gitignore.io/api/node
 # Edit at https://www.gitignore.io/?templates=node
