+++ src/client/github-api.json
@@ -1,4 +1,3 @@
 {
-  "uri": "https://api.github.com/graphql",
-  "token": "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"
+  "uri": "https://api.github.com/graphql"
 }
